Linux ASCII Compatibility
The Box-Drawing characters encoding is different on Linux and Windows.

Added generic definition at the beginning of "boardtools.c".
Replaced occurences of Drawwing-Box code in the code by generic values.

On Linux, the coding needs to be in String format. Therefore generic values are in String format, replaced %c by %s in the code.